+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Bug Fixes
 
+- **Fixed FOV-related sky stretching issues**
 - **Fixed weapon lowering under certain conditions**
 - **Fixed caching of certain NUGHUD patches** (fixed some crashes when using e.g. Status Bar graphics)
 - **Fixed vertical weapon inertia reset** when disabling mouselook/padlook